pyproject.toml
[project]
name = "agent-two"
version = "0.1.0"
description = "Multi-agent agentic app spec generator - Netanel Systems"
requires-python = ">=3.11"
dependencies = [
     "deepagents>=0.4,<0.5",
     "langchain-anthropic>=1.3,<2.0",
     "langchain-openai>=0.3,<1.0",
     "tavily-python>=0.7,<0.8",
     "python-dotenv>=1.0.0",
  ]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
  ]
